refactor CLI classes to own package `de.jplag.cli` moved LanguageLoader into CLI as only required there
