+++ cli/src/main/java/de/jplag/cli/LanguageLoader.java
@@ -1,4 +1,4 @@
-package de.jplag;
+package de.jplag.cli;
 
 import java.util.Collections;
 import java.util.Map;
@@ -10,6 +10,8 @@ import java.util.TreeSet;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import de.jplag.Language;
 
 /**
  * This class contains methods to load {@link Language Languages}.
