+++ cli/src/main/java/de/jplag/cli/logger/CollectedLogger.java
@@ -1,4 +1,4 @@
-package de.jplag.logger;
+package de.jplag.cli.logger;
 
 import java.io.PrintStream;
 import java.io.Serial;
