+++ cli/src/main/java/de/jplag/cli/logger/Triple.java
@@ -1,4 +1,4 @@
-package de.jplag.logger;
+package de.jplag.cli.logger;
 
 public record Triple<A, B, C> (A first, B second, C third) {
 }
