+++ cli/src/test/java/de/jplag/cli/ClusteringTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import de.jplag.CommandLineArgument;
 import de.jplag.clustering.Preprocessing;
 
 class ClusteringTest extends CommandLineInterfaceTest {
